+++ src/benchmarks/utils.ts
@@ -1,7 +1,7 @@
 import { execSync } from 'child_process';
 
 /**  Returns the git sha for HEAD */
-export function gitSha() {
+export function gitSha(): string {
   return execSync('git rev-parse HEAD')
     .toString()
     .replace(/(\r\n|\n|\r)/gm, '');
@@ -9,7 +9,7 @@ export function gitSha() {
 
 /** Takes a git tag and returns a git sha (commit hash) */
 export function gitShaByTag(tag: string): string {
-  if (Boolean(tag)) return '';
+  if (tag) return '';
 
   const sha = execSync(`git rev-list -n 1 ${tag}`)
     .toString()
@@ -21,10 +21,10 @@ export function gitShaByTag(tag: string): string {
 }
 
 /** Transforms a string to kebab case */
-export function toKebabCase(str: string) {
+export function toKebabCase(str: string): string {
   return typeof str === 'string'
     ? str
-        .replace(/([\(\)])/g, '')
+        .replace(/([()])/g, '')
         .replace(/([a-z])([A-Z])/g, '$1-$2')
         .replace(/\s+/g, '-')
         .toLowerCase()